+++ x/inflation/keeper/hooks.go
@@ -3,6 +3,8 @@ package keeper
 import (
 	"fmt"
 
+	"github.com/armon/go-metrics"
+	"github.com/cosmos/cosmos-sdk/telemetry"
 	sdk "github.com/cosmos/cosmos-sdk/types"
 	epochstypes "github.com/tharsis/evmos/v4/x/epochs/types"
 	"github.com/tharsis/evmos/v4/x/inflation/types"
@@ -12,7 +14,7 @@ import (
 func (k Keeper) BeforeEpochStart(_ sdk.Context, _ string, _ int64) {
 }
 
-// AfterEpochEnd mints and distributes coins at the end of each epoch end
+// AfterEpochEnd mints and allocates coins at the end of each epoch end
 func (k Keeper) AfterEpochEnd(ctx sdk.Context, epochIdentifier string, epochNumber int64) {
 	params := k.GetParams(ctx)
 	skippedEpochs := k.GetSkippedEpochs(ctx)
@@ -27,7 +29,7 @@ func (k Keeper) AfterEpochEnd(ctx sdk.Context, epochIdentifier string, epochNumb
 
 		k.SetSkippedEpochs(ctx, skippedEpochs)
 		k.Logger(ctx).Debug(
-			"skipping inflation mint and distribution",
+			"skipping inflation mint and allocation",
 			"height", ctx.BlockHeight(),
 			"epoch-id", epochIdentifier,
 			"epoch-number", epochNumber,
@@ -48,7 +50,8 @@ func (k Keeper) AfterEpochEnd(ctx sdk.Context, epochIdentifier string, epochNumb
 	}
 
 	mintedCoin := sdk.NewCoin(params.MintDenom, epochMintProvision.TruncateInt())
-	if err := k.MintAndAllocateInflation(ctx, mintedCoin); err != nil {
+	staking, incentives, communityPool, err := k.MintAndAllocateInflation(ctx, mintedCoin)
+	if err != nil {
 		panic(err)
 	}
 
@@ -79,6 +82,37 @@ func (k Keeper) AfterEpochEnd(ctx sdk.Context, epochIdentifier string, epochNumb
 		)
 		k.SetEpochMintProvision(ctx, newProvision)
 	}
+
+	defer func() {
+		if mintedCoin.Amount.IsInt64() {
+			telemetry.IncrCounterWithLabels(
+				[]string{types.ModuleName, "allocate", "total"},
+				float32(mintedCoin.Amount.Int64()),
+				[]metrics.Label{telemetry.NewLabel("denom", mintedCoin.Denom)},
+			)
+		}
+		if staking.AmountOf(mintedCoin.Denom).IsInt64() {
+			telemetry.IncrCounterWithLabels(
+				[]string{types.ModuleName, "allocate", "staking", "total"},
+				float32(staking.AmountOf(mintedCoin.Denom).Int64()),
+				[]metrics.Label{telemetry.NewLabel("denom", mintedCoin.Denom)},
+			)
+		}
+		if incentives.AmountOf(mintedCoin.Denom).IsInt64() {
+			telemetry.IncrCounterWithLabels(
+				[]string{types.ModuleName, "allocate", "incentives", "total"},
+				float32(incentives.AmountOf(mintedCoin.Denom).Int64()),
+				[]metrics.Label{telemetry.NewLabel("denom", mintedCoin.Denom)},
+			)
+		}
+		if communityPool.AmountOf(mintedCoin.Denom).IsInt64() {
+			telemetry.IncrCounterWithLabels(
+				[]string{types.ModuleName, "allocate", "community_pool", "total"},
+				float32(communityPool.AmountOf(mintedCoin.Denom).Int64()),
+				[]metrics.Label{telemetry.NewLabel("denom", mintedCoin.Denom)},
+			)
+		}
+	}()
 
 	ctx.EventManager().EmitEvent(
 		sdk.NewEvent(